Fix lint, and keep tmpdir when neither fix nor check given

+++ tasks/schema/generate.py
@@ -301,5 +301,5 @@ def codegen(ctx, keep_orig_order=False, check=False, fix=False, keeptmp=False):
         # Fix any differences by copying the codegen results into SETUP_INIT_DIR
         ctx.run(f"cp {tmpdir}/*_settings.go {SETUP_INIT_DIR}/")
 
-    if not keeptmp:
+    if not keeptmp and not display:
         shutil.rmtree(tmpdir)
